requirements.txt
google-api-python-client==2.86.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
yt-dlp>=2025.3.31
pyyaml==6.0.1
